Update EditToolView with new tool fields (limits, VICE, and OSG).
Updated some ToolContainer limits from Integers to Longs and Doubles.

+++ de-lib/src/main/java/org/iplantc/de/client/models/tool/ToolContainer.java
@@ -25,16 +25,16 @@ public interface ToolContainer extends HasName {
     void setEntryPoint(String entryPoint);
 
     @PropertyName("memory_limit")
-    Integer getMemoryLimit();
+    Long getMemoryLimit();
 
     @PropertyName("memory_limit")
-    void setMemoryLimit(Integer memoryLimit);
+    void setMemoryLimit(Long memoryLimit);
 
     @PropertyName("min_memory_limit")
-    Integer getMinMemoryLimit();
+    Long getMinMemoryLimit();
 
     @PropertyName("min_memory_limit")
-    void setMinMemoryLimit(Integer memoryLimit);
+    void setMinMemoryLimit(Long memoryLimit);
 
     @PropertyName("cpu_shares")
     Integer getCpuShares();
@@ -79,16 +79,16 @@ public interface ToolContainer extends HasName {
     Integer getPidsLimit();
 
     @PropertyName("max_cpu_cores")
-    Integer getMaxCPUCores();
+    Double getMaxCPUCores();
 
     @PropertyName("max_cpu_cores")
-    void setMaxCPUCores(Integer cores);
+    void setMaxCPUCores(Double cores);
 
     @PropertyName("min_cpu_cores")
-    Integer getMinCPUCores();
+    Double getMinCPUCores();
 
     @PropertyName("min_cpu_cores")
-    void setMinCPUCores(Integer cores);
+    void setMinCPUCores(Double cores);
 
     @PropertyName("skip_tmp_mount")
     Boolean isSkipTmpMount();
@@ -97,10 +97,10 @@ public interface ToolContainer extends HasName {
     void setSkipTmpMount(Boolean skip);
 
     @PropertyName("min_disk_space")
-    Integer getMinDiskSpace();
+    Long getMinDiskSpace();
 
     @PropertyName("min_disk_space")
-    void setMinDiskSpace(Integer diskSpace);
+    void setMinDiskSpace(Long diskSpace);
 
     @PropertyName("container_ports")
     List<ToolContainerPort> getContainerPorts();
